feat: refact validator in another file
